renamed routes in Controllers made for private use
